# Fix trigger for tier 1 call counting delay