--- a/src/vm/prestub.cpp
+++ b/src/vm/prestub.cpp
@@ -734,12 +734,10 @@ PCODE MethodDesc::JitCompileCodeLockedEventWrapper(PrepareCodeConfig* pConfig, J
     }
 
 #ifdef FEATURE_TIERED_COMPILATION
-    if (flags.IsSet(CORJIT_FLAGS::CORJIT_FLAG_TIER0))
+    if (g_pConfig->TieredCompilation() && flags.IsSet(CORJIT_FLAGS::CORJIT_FLAG_TIER0))
     {
-        // The flag above is only set (in TieredCompilationManager::GetJitFlags()) when tiered compilation is enabled, this
-        // method was eligible for tiered compilation at the time when it was checked, and a tier 0 JIT was requested for this
-        // method
-        _ASSERTE(g_pConfig->TieredCompilation());
+        // The flag above is only set (in TieredCompilationManager::GetJitFlags()) when this method was eligible for tiered
+        // compilation at the time when it was checked, and a tier 0 JIT was requested for this method
         GetAppDomain()->GetTieredCompilationManager()->OnTier0JitInvoked();
     }
 #endif // FEATURE_TIERED_COMPILATION