@kouvel 2018-04-09
```
Before:

       Benchmark                Metric                Default                 Tiering                 Minopts
-----------------------  --------------------  ----------------------  ----------------------  ----------------------
Dotnet_Build_HelloWorld         Duration (ms)        2002 (2002-2002)        1560 (1560-1560)        1523 (1523-1523)
        Csc_Hello_World         Duration (ms)           872 (872-872)           589 (589-589)           576 (576-576)
      Csc_Roslyn_Source         Duration (ms)        2730 (2730-2730)        2663 (2663-2663)        2500 (2500-2500)
               Word2Vec         Training (ms)  365666 (365666-365666)  390065 (390065-390065)  543122 (543122-543122)
               Word2Vec     First Search (ms)              70 (70-70)           310 (310-310)           348 (348-348)
               Word2Vec    Median Search (ms)              67 (67-67)              69 (69-69)           347 (347-347)
             MusicStore          Startup (ms)           981 (981-981)           718 (718-718)           678 (678-678)
             MusicStore    First Request (ms)        1096 (1096-1096)           665 (665-665)           638 (638-638)
             MusicStore  Median Response (ms)                 3 (3-3)                 2 (2-2)                 3 (3-3)
               AllReady          Startup (ms)        1798 (1798-1798)        1385 (1385-1385)        1297 (1297-1297)
               AllReady    First Request (ms)           685 (685-685)           446 (446-446)           415 (415-415)
               AllReady  Median Response (ms)                 4 (4-4)                 5 (5-5)                 4 (4-4)

After:

       Benchmark                Metric                Default                 Tiering                 Minopts
-----------------------  --------------------  ----------------------  ----------------------  ----------------------
Dotnet_Build_HelloWorld         Duration (ms)        2016 (2016-2016)        1566 (1566-1566)        1527 (1527-1527)
        Csc_Hello_World         Duration (ms)           871 (871-871)           590 (590-590)           574 (574-574)
      Csc_Roslyn_Source         Duration (ms)        2771 (2771-2771)        2640 (2640-2640)        2513 (2513-2513)
               Word2Vec         Training (ms)  366437 (366437-366437)  388025 (388025-388025)  538938 (538938-538938)
               Word2Vec     First Search (ms)              71 (71-71)           309 (309-309)           352 (352-352)
               Word2Vec    Median Search (ms)              67 (67-67)              69 (69-69)           347 (347-347)
             MusicStore          Startup (ms)           985 (985-985)           712 (712-712)           684 (684-684)
             MusicStore    First Request (ms)        1077 (1077-1077)           661 (661-661)           633 (633-633)
             MusicStore  Median Response (ms)                 3 (3-3)                 2 (2-2)                 3 (3-3)
               AllReady          Startup (ms)        1802 (1802-1802)        1374 (1374-1374)        1287 (1287-1287)
               AllReady    First Request (ms)           687 (687-687)           457 (457-457)           413 (413-413)
               AllReady  Median Response (ms)                 4 (4-4)                 3 (3-3)                 4 (4-4)
```

MusicStore

```
No tiering:

Server start (ms):   992
1st Request (ms):   1082
Total (ms):         2074

  Requests    Aggregate Time(ms)    Req/s   Req Min(ms)   Req Mean(ms)   Req Median(ms)   Req Max(ms)   SEM(%)
-----------   ------------------   ------   -----------   ------------   --------------   -----------   ------
    2-  100                 2409   294.85          2.80           3.39             3.02         31.01     8.35
  101-  250                 2861   332.10          2.57           3.01             3.00          3.89     0.66
  251-  500                 3538   369.03          2.28           2.71             2.60         10.51     1.31
  501-  750                 4215   369.67          2.34           2.71             2.60          3.72     0.62
  751- 1000                 4909   360.32          2.29           2.78             2.85          3.48     0.50
 1001- 1500                 6308   357.24          2.12           2.80             2.83          8.64     0.64
 1501- 2000                 7671   366.79          2.29           2.73             2.78          8.14     0.53
 2001- 3000                10430   362.49          2.22           2.76             2.78          7.93     0.29
 3001- 5000                15902   365.47          2.24           2.74             2.78          4.22     0.16
 5001-10000                30162   350.63          2.11           2.85             2.85          7.15     0.15

Tiering (before):

Server start (ms):   698
1st Request (ms):    666
Total (ms):         1364

  Requests    Aggregate Time(ms)    Req/s   Req Min(ms)   Req Mean(ms)   Req Median(ms)   Req Max(ms)   SEM(%)
-----------   ------------------   ------   -----------   ------------   --------------   -----------   ------
    2-  100                 1800   226.99          2.96           4.41             3.61         21.13     5.31
  101-  250                 2237   343.19          2.63           2.91             2.85          4.04     0.71
  251-  500                 2856   403.95          1.98           2.48             2.41         10.34     1.47
  501-  750                 3423   440.47          1.86           2.27             2.19          3.33     0.70
  751- 1000                 3991   439.99          2.08           2.27             2.22          3.19     0.54
 1001- 1500                 5149   432.02          1.95           2.31             2.29          8.20     0.65
 1501- 2000                 6297   435.28          1.47           2.30             2.19         10.38     0.97
 2001- 3000                 8515   451.02          1.68           2.22             2.15          3.39     0.37
 3001- 5000                12873   458.93          1.50           2.18             2.13          7.27     0.29
 5001-10000                24298   437.63          1.52           2.29             2.33          6.08     0.16

Tiering (after):

Server start (ms):   710
1st Request (ms):    668
Total (ms):         1378

  Requests    Aggregate Time(ms)    Req/s   Req Min(ms)   Req Mean(ms)   Req Median(ms)   Req Max(ms)   SEM(%)
-----------   ------------------   ------   -----------   ------------   --------------   -----------   ------
    2-  100                 1811   228.51          2.84           4.38             3.68         20.97     5.30
  101-  250                 2247   343.62          2.66           2.91             2.85          4.04     0.68
  251-  500                 2890   388.82          2.06           2.57             2.55         12.05     1.63
  501-  750                 3443   452.02          1.83           2.21             2.15          3.32     0.64
  751- 1000                 3995   453.21          1.87           2.21             2.17          3.22     0.63
 1001- 1500                 5102   451.53          1.67           2.21             2.18          9.59     0.83
 1501- 2000                 6151   476.71          1.86           2.10             2.07          3.03     0.35
 2001- 3000                 8328   459.45          1.56           2.18             2.15          7.58     0.47
 3001- 5000                12806   446.63          1.63           2.24             2.32          3.80     0.25
 5001-10000                24161   440.30          1.61           2.27             2.32          9.70     0.17
```

AllReady

```
No tiering:

Server start (ms):  1820
1st Request (ms):    702
Total (ms):         2522

  Requests    Aggregate Time(ms)    Req/s   Req Min(ms)   Req Mean(ms)   Req Median(ms)   Req Max(ms)   SEM(%)
-----------   ------------------   ------   -----------   ------------   --------------   -----------   ------
    2-  100                 2919   249.04          3.52           4.02             3.92          9.45     1.61
  101-  250                 3534   243.92          3.59           4.10             4.00         17.38     2.23
  251-  500                 4482   263.77          3.44           3.79             3.68         11.42     0.99
  501-  750                 5416   267.62          3.41           3.74             3.65         14.03     1.14
  751- 1000                 6343   269.72          3.41           3.71             3.64         11.05     0.85
 1001- 1500                 8178   272.48          3.34           3.67             3.63          5.23     0.23
 1501- 2000                10049   267.15          3.34           3.74             3.66          5.60     0.37

Tiering (before):

Server start (ms):  1407
1st Request (ms):    464
Total (ms):         1871

  Requests    Aggregate Time(ms)    Req/s   Req Min(ms)   Req Mean(ms)   Req Median(ms)   Req Max(ms)   SEM(%)
-----------   ------------------   ------   -----------   ------------   --------------   -----------   ------
    2-  100                 2376   195.98          4.48           5.10             4.99         11.91     1.47
  101-  250                 3100   207.17          4.28           4.83             4.66         17.01     1.77
  251-  500                 4273   213.05          4.23           4.69             4.58         14.17     0.93
  501-  750                 5434   215.30          4.32           4.64             4.56         12.24     0.71
  751- 1000                 6579   218.38          4.20           4.58             4.55          5.70     0.25
 1001- 1500                 8885   216.85          4.22           4.61             4.57          6.54     0.20
 1501- 2000                11195   216.44          4.16           4.62             4.58          6.16     0.19

Tiering (after):

Server start (ms):  1400
1st Request (ms):    464
Total (ms):         1864

  Requests    Aggregate Time(ms)    Req/s   Req Min(ms)   Req Mean(ms)   Req Median(ms)   Req Max(ms)   SEM(%)
-----------   ------------------   ------   -----------   ------------   --------------   -----------   ------
    2-  100                 2395   186.43          3.86           5.36             4.58         19.38     4.50
  101-  250                 3004   246.26          3.51           4.06             4.01          5.58     0.69
  251-  500                 3868   289.17          3.05           3.46             3.37         17.64     1.75
  501-  750                 4665   313.91          3.00           3.19             3.12          5.86     0.46
  751- 1000                 5453   316.94          2.81           3.16             3.10          4.47     0.39
 1001- 1500                 7029   317.33          2.82           3.15             3.12          4.60     0.25
 1501- 2000                 8602   317.85          2.78           3.15             3.12          4.42     0.28
```